% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_complementary_features.R
\name{aggregate_complementary_features}
\alias{aggregate_complementary_features}
\title{Aggreggate complementary features}
\usage{
aggregate_complementary_features(X, rt_class = NULL)
}
\description{
Aggreggate complementary features
}
