% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rt.R
\name{align_rt}
\alias{align_rt}
\title{Align metabolites from different batches on the basis of retention times}
\usage{
align_rt(X = NULL, rt_err = 1, rt_class = NULL)
}
\arguments{
\item{X}{data.frame with the following columns: PeakName, RT}

\item{rt_err}{maximum time difference between retention times of the same metabolite}

\item{rt_class}{internal parameter set by default to rt_err/2}
}
\description{
Align metabolites from different batches on the basis of retention times
}
