% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_univariate_results.R
\name{annotate_univariate_results}
\alias{annotate_univariate_results}
\title{Get significant features}
\usage{
annotate_univariate_results(
  mRList = NULL,
  feature_stats = NULL,
  add.metabolite.levels = TRUE,
  dirout = NULL
)
}
\arguments{
\item{mRList}{mRList object}

\item{feature_stats}{name of an element of mRList object containing results of statistical test or a custom data frame with Feature_ID as row names.}

\item{add.metabolite.levels}{if TRUE, adds mRList$data_ann, if this exists.}

\item{dirout}{output directory}
}
\value{
a data.frame
}
\description{
Merge mRList$metab_ann with feature_stats
}
