% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_dataset.R
\name{annotation_dataset}
\alias{annotation_dataset}
\title{generate the data set for annotation}
\usage{
annotation_dataset(
  column = c("HILIC", "LipC8", "pZIC", "RPLong", "RPShort"),
  mode = c("POS", "NEG"),
  accept_RI = 10
)
}
\description{
generate the data set for annotation
}
