% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.metaboset.mRList.R
\name{as.metaboset.mRList}
\alias{as.metaboset.mRList}
\title{Transforms mRList into mSet object}
\usage{
as.metaboset.mRList(mRList)
}
\arguments{
\item{mRList}{mRList object}
}
\value{
mset object
}
\description{
Transforms mRList into mSet object
}
