% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.metaboset.R
\name{as.metaboset}
\alias{as.metaboset}
\title{Transforms mRList into MetaboSet object}
\usage{
as.metaboset(mRList = NULL)
}
\arguments{
\item{mRList}{mRList object}
}
\value{
metabosets
}
\description{
Transforms mRList into MetaboSet object
}
