% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.MSnSet.mRList.R
\name{as.MSnSet.mRList}
\alias{as.MSnSet.mRList}
\title{Transforms mRList into MSnSet object}
\usage{
as.MSnSet.mRList(mRList)
}
\arguments{
\item{mRList}{mRList object}
}
\value{
MSnSet object
}
\description{
Transforms mRList into MSnSet object
}
