% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_ora.R
\name{barplot_ora}
\alias{barplot_ora}
\title{Draw a bar plot from pathway analyses with ORA}
\usage{
barplot_ora(
  pa_res,
  p = c("pvalue", "p.adjust"),
  filename = "pathway_analysis_ora"
)
}
\arguments{
\item{pa_res}{pa_res obtained from pathway_analysis with type "ora".}

\item{p}{"pvalue" or "p.adjust".}

\item{filename}{Name of the .png file that will be created.}
}
\description{
Draw a bar plot from pathway analyses with ORA
}
