% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsid2cid.R
\docType{data}
\name{bsid2cid}
\alias{bsid2cid}
\title{Biosystems ID to CID map}
\format{
An object of class \code{data.frame} with 8030771 rows and 3 columns.
}
\usage{
bsid2cid
}
\description{
Biosystems ID to CID map
}
\keyword{datasets}
