% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsid2info.R
\docType{data}
\name{bsid2info}
\alias{bsid2info}
\title{Biosystems ID annotation}
\format{
An object of class \code{data.frame} with 591873 rows and 8 columns.
}
\usage{
bsid2info
}
\description{
Biosystems ID annotation
}
\keyword{datasets}
