% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsid2pubchem.R
\docType{data}
\name{bsid2pubchem}
\alias{bsid2pubchem}
\title{Biosystems ID annotation}
\format{
An object of class \code{data.frame} with 663976 rows and 8 columns.
}
\usage{
bsid2pubchem
}
\description{
Biosystems ID annotation
}
\keyword{datasets}
