% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ppm_err.R
\name{calc_ppm_err}
\alias{calc_ppm_err}
\title{Calculate PPM error}
\usage{
calc_ppm_err(a = NULL, b = NULL)
}
\arguments{
\item{a}{data.frame of 1 row and PPM in column 1}

\item{b}{data.frame of 1 row and PPM in column 1}
}
\value{
PPM error
}
\description{
PPM_error = abs(x - y) / x * 1000000
}
