% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_reference.R
\name{calc_reference}
\alias{calc_reference}
\title{Calculate a reference profile}
\usage{
calc_reference(
  mRList = NULL,
  sample_col = "class",
  sample_class = "QC",
  approach = c("median", "mean")
)
}
\arguments{
\item{mRList}{mRList object}

\item{sample_col}{the column of mRList$sample_ann where the "sample_class" value is found}

\item{sample_class}{label to identify the samples to be used for the reference profile calculation}

\item{approach}{The type of average: "median" or "mean".}
}
\value{
mRList object with mRList$reference
}
\description{
The reference profile is calculated using the samples annotated as "sample_class" in the sample annotation column "sample_col".
Launch function before normalize_profiles function
}
\examples{
##library(dataset.margheRita)
##dataset(norm_pos)
mRList<-calc_reference(mRList=mRList, sample_col="class", sample_class="QC", approach="mean")
}
