% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RI.R
\name{calc_RI}
\alias{calc_RI}
\title{Convert sample intensity to relative intensity and filter}
\usage{
calc_RI(feature_spectra = NULL, accept_RI = 10)
}
\arguments{
\item{feature_spectra}{it is a list of sample ID which each contains list of its m/z and Intensity.}

\item{accept_RI}{numeric. the default value is 10. it is a maximum relative intensity that is kept in sample. since low intense peaks could be noise, it is filtering sample dataset by deleting the relative intensity lower then accept_RI.}
}
\value{
list of spectra with RI
}
\description{
Convert sample intensity to relative intensity and filter
}
