% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lfc.R
\name{calculate_lfc}
\alias{calculate_lfc}
\title{Calculate the log fold change between two classes of samples}
\usage{
calculate_lfc(mRList, contrast_samples)
}
\arguments{
\item{mRList}{mRList object}

\item{contrast_samples}{character vector with two elements that indicate the two classes of samples}
}
\value{
mRList object with fold changes in mRList$metab_ann
}
\description{
The log2 fold changes are calculated using the mean values of the two classes
}
