% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lfc_all.R
\name{calculate_lfc_all}
\alias{calculate_lfc_all}
\title{Calculate the log fold change between all pairs of sample classes}
\usage{
calculate_lfc_all(mRList)
}
\arguments{
\item{mRList}{mRList object}
}
\value{
mRList object with fold changes in mRList$metab_ann
}
\description{
Calculate the log fold change between all pairs of sample classes
}
\examples{
##library(dataset.margheRita)
##dataset(norm_pos)
mRList<-calculate_lfc_all(mRList)
}
