% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_annotation.R
\name{check_annotation}
\alias{check_annotation}
\title{Metabolite Annotation}
\usage{
check_annotation(
  feature_data = NULL,
  reference = NULL,
  feature_data_spectra = NULL,
  reference_spectra = NULL
)
}
\description{
Metabolite Annotation
}
