% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intense_peak.R
\name{check_intense_peak}
\alias{check_intense_peak}
\title{check the ppm error for the intense peaks}
\usage{
check_intense_peak(
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  n_peaks = 1,
  acceptable_PPM_err = 10
)
}
\description{
check the ppm error for the intense peaks
}
