% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intense_peak2.R
\name{check_intense_peak2}
\alias{check_intense_peak2}
\title{Check the ppm error of MS/MS peaks}
\usage{
check_intense_peak2(
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  reference = NULL,
  lib_peaks_data = NULL,
  n_peaks = 1,
  acceptable_PPM_err = 10,
  mode = c("POS", "NEG")
)
}
\arguments{
\item{RT_mass}{a list of data frame, obtained by check_RT_mass() function. it is contains a list of metabolite with desired retention time and PPM error. each metabolite is a data frame of potential candidates with specific features; Feature_ID, RT_err, RT_flag, ppm_error, mass_status and mass_flag.}

\item{RI_lib}{it is generated by RI_lib_data() or select_library() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks. note that in house library of the margheRita is already stored by relative intensity and filtered by select_library() function so, in this case RI_lib is lib_peaks which is generated by select_library() function.}

\item{RI_sample}{it is generated by RI_sample_data() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{reference}{A list of metabolite with their specific features of retention time and mz which stored by their ID, CAS number and Name.}

\item{lib_peaks_data}{it is generated by select_library() function. it is providing information regarding the peaks such as ID, CAS number, Name and POS/NEG collision energy.}

\item{n_peaks}{A number with default value of 1. check_intense_peak() function calculate the PPM error peak by peak starting from the most intense one. n_peaks is the number of the peaks which will be calculated for each metabolite and candidate.}

\item{acceptable_PPM_err}{A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{mode}{mode could be set in positive or negative state. positive mode select positive collision energy and mz in positive mode.}
}
\value{
a data frame with metabolite annotation
}
\description{
Check the ppm error of MS/MS peaks
}
