% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intense_peak3.R
\name{check_intense_peak3}
\alias{check_intense_peak3}
\title{peak matching}
\usage{
check_intense_peak3(
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  reference = NULL,
  lib_peaks_data = NULL,
  ppm_err = 10,
  intensity = 30,
  mode = c("POS", "NEG")
)
}
\description{
peak matching
}
