% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mass.R
\name{check_mass}
\alias{check_mass}
\title{check_mass}
\usage{
check_mass(
  feature_data,
  reference,
  unaccept_flag = unaccept_flag,
  accept_flag = accept_flag,
  suffer_flag = suffer_flag
)
}
\description{
check_mass
}
