% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mass_mass.R
\name{check_mass_mass}
\alias{check_mass_mass}
\title{calculating and selecting proper PPM error and RI for mass mass}
\usage{
check_mass_mass(RT_mass, RI_lib, RI_sample)
}
\description{
calculating and selecting proper PPM error and RI for mass mass
}
