% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_ratio.R
\name{CV_ratio}
\alias{CV_ratio}
\title{Coefficient of variation filter}
\usage{
CV_ratio(mRList = NULL, dirout = NULL, ratioCV = 1)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}

\item{ratioCV}{only metabolites with CV ratio (sample/QC) higher than ratioCV will be kept}
}
\value{
filtered mRList object
}
\description{
Calculate the coefficient of variance (CV) for samples and QC (quality control) to assess the reliability of the experiment, as:
CV = (SD - mean) * 100.
CV samples and CV QC are compared, calculating the ratio between CV samples and CV QC.
The threshold of the ratio has to be decided by the user in ratioCV.
Only metabolites with ratio >= ratioCV are used for the analysis.
}
