% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metabolite_associations.R
\name{filter_metabolite_associations}
\alias{filter_metabolite_associations}
\title{Filter metabolite-feature pairs}
\usage{
filter_metabolite_associations(mRList = NULL, strategy = c("nrpl", "lpn"))
}
\arguments{
\item{mRList}{mRList object}

\item{strategy}{"nrpl" stands for number of peaks, peak ratio, ppm error, level; "lpn" stands for level, ppm error, number of peaks}
}
\description{
Filter metabolite-feature pairs
}
