% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering of features}
\usage{
filtering(
  mRList = NULL,
  seed = NULL,
  a = 0.1,
  b = 0.25,
  lower_quality_mass_acc = 0.4,
  upper_quality_mass_acc = 0.8,
  min_metab_in_sample = 100,
  min_sample_with_metab = 3,
  na_value = "NA"
)
}
\arguments{
\item{mRList}{mRList object}

\item{seed}{set this seed with set.seed}

\item{a}{minimum increase factor}

\item{b}{maxmimum increase factor}

\item{lower_quality_mass_acc}{lower quality value}

\item{upper_quality_mass_acc}{upper quality value}

\item{min_metab_in_sample}{min number of metabolites in a sample}

\item{min_sample_with_metab}{min number of samples in which a metabolite must appear}

\item{na_value}{value that indicate missing values}
}
\value{
filtered mRList object
}
\description{
Filtering of features by missing values, m/z values and imputation of missing values
}
