% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset_for_annotation.R
\name{generate_dataset_for_annotation}
\alias{generate_dataset_for_annotation}
\title{margherRita - main function to run the full pipeline}
\usage{
generate_dataset_for_annotation(wdir = "./")
}
\description{
margherRita - main function to run the full pipeline
}
