% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectra_list_from_vector.R
\name{get_spectra_list_from_vector}
\alias{get_spectra_list_from_vector}
\title{get_spectra_list_from_vector}
\usage{
get_spectra_list_from_vector(spectra)
}
\description{
get_spectra_list_from_vector
}
