% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_map.2.R
\name{h_map}
\alias{h_map}
\title{heatmap}
\usage{
h_map(
  mRList,
  dirout = "./",
  col_ann = NULL,
  col = NULL,
  scale_features = TRUE,
  features = NULL,
  top = 500,
  ...
)
}
\arguments{
\item{mRList}{mRList}
}
\description{
heatmap
}
