% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_map_MSMS_comparison.R
\name{h_map_MSMS_comparison}
\alias{h_map_MSMS_comparison}
\title{Vh_map_MSMS_comparison}
\usage{
h_map_MSMS_comparison(
  matrices_list = NULL,
  out_prefix = "MSMS_comparison",
  col = NULL,
  ppm_error = 10,
  RI_difference = 10
)
}
\arguments{
\item{matrices_list}{list of matrices with ppm error, RI difference and flags}

\item{out_prefix}{prefix for output files}
}
\description{
Vh_map_MSMS_comparison
}
