% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatscatter_chromatography.R
\name{heatscatter_chromatography}
\alias{heatscatter_chromatography}
\title{Heatscatter chromatography plot
Draw the "heatscatter chromatography" and save it to a png file}
\usage{
heatscatter_chromatography(
  mRList = NULL,
  mz_limits = NULL,
  rt_limits = NULL,
  sample = NULL,
  out_file = "heatscatter_chromatography.png",
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{mz_limits}{numeric vector with minimum and maximum rt values}

\item{rt_limits}{numeric vector with minimum and maximum rt values}

\item{sample}{if not null, consider only the given samples}

\item{out_file}{png file name}

\item{...}{further arguments to LSD::heatscatter}
}
\description{
Heatscatter chromatography plot
Draw the "heatscatter chromatography" and save it to a png file
}
