% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_strongest_representative.R
\name{keep_strongest_representative}
\alias{keep_strongest_representative}
\title{Keep the feature with the highest average across samples}
\usage{
keep_strongest_representative(X = NULL)
}
\arguments{
\item{X}{features-by-samples data matrix}
}
\description{
Keep the feature with the highest average across samples
}
