% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nutrition_dataset.R
\name{load_nutrition_dataset}
\alias{load_nutrition_dataset}
\title{Load nutrition dataset}
\usage{
load_nutrition_dataset(column = NULL, normalized = FALSE)
}
\arguments{
\item{column}{POS or NEG}

\item{normalized}{dataset with/without normalized samples.#'}
}
\description{
Load nutrition dataset
}
