% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margheRita_library.R
\name{margheRita_library}
\alias{margheRita_library}
\title{Retrieve margheRita metabolite level 2 annotation library}
\usage{
margheRita_library(
  column = c("HILIC", "LipC8", "pZIC", "RPLong", "RPShort"),
  mode = c("POS", "NEG"),
  accept_RI = 10
)
}
\arguments{
\item{column}{it is specify the type of the column which could be; HILIC, LipC8, pZIC, RPLong, RPShort. based on the type of the column, the retention time of same metabolite could be different.}

\item{mode}{mode could be set in positive (POS) or negative (NEG). positive mode select positive collision energy and mz in positive mod.}

\item{accept_RI}{numeric. the default value is 10. it is a maximum relative intensity that we keep in library. since low intense peaks could be noise, the library is filtered and only peaks with relative intesity > accept_RI are kept.}
}
\value{
a list with three elements:
\itemize{
\item precursors library: a list of metabolites which each contains information of retention time and mz with their specific IDs, CAS numbers and names.
\item lib_peaks_data: a list of metabolites which each contains information of collision energy with their specific IDs, CAS numbers and names.
\item lib_peaks: a list of metabolites with same IDs as lib_peaks_data which each contains a list of a peaks with mz and relative intensity.
}
}
\description{
Retrieve margheRita metabolite level 2 annotation library
}
