% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margheRita_test.R
\name{margheRita_test}
\alias{margheRita_test}
\title{margherRita - main function to run the full pipeline}
\usage{
margheRita_test(wdir = "./")
}
\description{
margherRita - main function to run the full pipeline
}
