% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margheRita_test_compare_pqn_driftcorr.R
\name{margheRita_test_compare_pqn_driftcorr}
\alias{margheRita_test_compare_pqn_driftcorr}
\title{margherRita - main function to run the full pipeline}
\usage{
margheRita_test_compare_pqn_driftcorr(wdir = "./")
}
\description{
margherRita - main function to run the full pipeline
}
