% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margheRita_test_metab_annotation.R
\name{margheRita_test_metab_annotation}
\alias{margheRita_test_metab_annotation}
\title{a fake function to test pieces of code}
\usage{
margheRita_test_metab_annotation(wdir = "./")
}
\description{
a fake function to test pieces of code
}
