% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margheRita_test_pa.R
\name{margheRita_test_pa}
\alias{margheRita_test_pa}
\title{margherRita - main function to run the full pipeline}
\usage{
margheRita_test_pa(wdir = "./")
}
\description{
margherRita - main function to run the full pipeline
}
