% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median_stdev_samples.R
\name{mean_media_stdev_samples}
\alias{mean_media_stdev_samples}
\title{mean, median, and stdev for technical replicates of each biological sample}
\usage{
mean_media_stdev_samples(dataframe_post, metadata)
}
\arguments{
\item{metadata}{split in metadata for QC and metadata for samples}

\item{calculate}{mean, median and stdev for samples}
}
\description{
mean, median, and stdev for technical replicates of each biological sample
}
