% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median_stdev_samples.R
\name{mean_median_stdev_samples}
\alias{mean_median_stdev_samples}
\title{Calculate mean, median and standard deviation (stdev) for each group under study.
write_output =FALSE default, if TRUE table, as .csv format, are written}
\usage{
mean_median_stdev_samples(mRList, dirout = "./", write_output = FALSE)
}
\arguments{
\item{mRList}{mRList}
}
\description{
Calculate mean, median and standard deviation (stdev) for each group under study.
write_output =FALSE default, if TRUE table, as .csv format, are written
}
