% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_stats_w_features.R
\name{merge_stats_w_features}
\alias{merge_stats_w_features}
\title{Merge feature files with statistics}
\usage{
merge_stats_w_features(
  mRList = NULL,
  test_method = "anova",
  test_value = "q",
  cutoff_value = 0.05
)
}
\arguments{
\item{mRList}{mRList object}

\item{test_method}{name of an element of mRList object containing results of statistical test or a custom data frame with Feature_ID as row names (e.g.: "anova").}

\item{test_value}{value to use for significance. e.g.: "p" or "q"}

\item{cutoff_value}{value below which a feature is considered significant.}
}
\value{
the data.frame metab_stat, metab_stat_w_features, metab_stat_w_features_known, metab_stat_w_features_only_sign, metab_stat_w_features_known_only_sign will be both saved in the global environemtn and exported as txt tab delimited files.
}
\description{
Merge feature files with results of the statistics, and export the results
}
