% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_annotation.R
\name{metabolite_annotation}
\alias{metabolite_annotation}
\title{metabolite_annotation
map between metabolite name and various stantard identifiers}
\usage{
metabolite_annotation(mRList = NULL, reference = NULL)
}
\arguments{
\item{mRList}{mRList object}

\item{reference}{data.frame with columns "rt", "mz", "Name", "MS_MS_spectrum"}
}
\value{
the annotation list, which has three elements... 1) named list of data.frames; each element of the list corresponds to an element of reference and each data.frame containes the candidates
}
\description{
metabolite_annotation
map between metabolite name and various stantard identifiers
}
