% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_annotation2.R
\name{metabolite_annotation2}
\alias{metabolite_annotation2}
\title{Metabolite Annotation}
\usage{
metabolite_annotation2(
  reference = NULL,
  feature_data = NULL,
  RI_lib = NULL,
  lib_peaks_data = NULL,
  RI_sample = NULL,
  rt_err_thr = 1,
  unaccept_flag = 15,
  accept_flag = 5,
  suffer_flag = 10,
  mode = c("NEG", "POS"),
  n_peaks = 1,
  acceptable_PPM_err = 10
)
}
\description{
Metabolite Annotation
}
