% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_annotation 3.R
\name{metabolite_annotation3}
\alias{metabolite_annotation3}
\title{Metabolite Annotation}
\usage{
metabolite_annotation3(
  feature_data = NULL,
  reference = NULL,
  feature_spectra = NULL,
  reference_spectra = NULL,
  rt_err_thr = 1,
  unaccept_flag = 15,
  accept_flag = 5,
  suffer_flag = 10,
  acceptable_RI = 10,
  n_peaks = 1,
  acceptable_PPM_err = 10
)
}
\description{
Metabolite Annotation
}
