% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_family.R
\name{metabolite_family}
\alias{metabolite_family}
\title{family of metabolites}
\usage{
metabolite_family(
  reference = NULL,
  feature_data = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  lib_peaks_data = NULL,
  rt_err_thr = 1,
  unaccept_flag = 10,
  mode = c("NEG", "POS"),
  ppm_err = 10,
  intensity = 30
)
}
\description{
family of metabolites
}
