% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mL_demo.R
\docType{data}
\name{mL_demo}
\alias{mL_demo}
\title{mL_demo}
\format{
An object of class \code{list} of length 12.
}
\usage{
mL_demo
}
\description{
data structure that contains data for the vignettes
}
\keyword{datasets}
