% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRlib_peaks_df.R
\docType{data}
\name{mRlib_peaks_df}
\alias{mRlib_peaks_df}
\title{LIBRARY - List of peaks as data.frame}
\format{
An object of class \code{data.frame} with 2132 rows and 9 columns.
}
\usage{
mRlib_peaks_df
}
\description{
LIBRARY - List of peaks as data.frame
}
\keyword{datasets}
