% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRlib_peaks_list.R
\docType{data}
\name{mRlib_peaks_list}
\alias{mRlib_peaks_list}
\title{LIBRARY - List of peaks}
\format{
An object of class \code{list} of length 2154.
}
\usage{
mRlib_peaks_list
}
\description{
LIBRARY - List of peaks
}
\keyword{datasets}
