% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRlib_precursors.R
\docType{data}
\name{mRlib_precursors}
\alias{mRlib_precursors}
\title{LIBRARY - data.frame wirh information at precursor level}
\format{
An object of class \code{data.frame} with 1871 rows and 16 columns.
}
\usage{
mRlib_precursors
}
\description{
LIBRARY - data.frame wirh information at precursor level
}
\keyword{datasets}
