% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_check.R
\name{norm_check}
\alias{norm_check}
\title{Test the normality of each metabolite by Shapiro-Wilk's test}
\usage{
norm_check(mRList)
}
\arguments{
\item{mRList}{mRList object}
}
\value{
mRList object with mRist$Shapiro elment that reports p-value calculated using Shapiro-Wilk's test
}
\description{
Function calculates p-values for each metabolites by Shapiro-Wilk's test to test normal distribution. If p-value>0.05 metabolite is normally distributed, p-value<0.05 metabolite is not normally distributed
}
