% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_profiles.R
\name{normalize_profiles}
\alias{normalize_profiles}
\title{Normalize profiles}
\usage{
normalize_profiles(mRList = NULL, method = c("log", "reference", "pqn"))
}
\arguments{
\item{mRList}{mRList object}

\item{method}{"log": log2(x+1); "reference": divide each sample by its reference; "pqn": probailistic quotient normalization;}
}
\value{
mRList object with normalized data
}
\description{
Normalize profiles
}
