% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{Pathway analysis
Run enricher or gsea from clusterProfiler on pathways provided by NCBI Biosytems}
\usage{
pathway_analysis(
  in_list = NULL,
  universe = NULL,
  type = c("ora", "gsea"),
  tax_id = 9606,
  include_general_pathways = FALSE,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  TERM2METAB = NULL,
  TERM2NAME = NULL,
  nPerm = 1000,
  verbose = TRUE,
  gsea_by = "fgsea",
  seed = FALSE,
  exponent = 1
)
}
\arguments{
\item{in_list}{metabolite set}

\item{universe}{universe}

\item{type}{"ora" or "gsea"}

\item{tax_id}{tax id}

\item{include_general_pathways}{whether to include general pathways or not}

\item{pvalueCutoff}{p-value cutoff}

\item{pAdjustMethod}{p-value adjustment method}

\item{minGSSize}{minimum metabolite set size}

\item{maxGSSize}{maximum metabolite set size}

\item{qvalueCutoff}{q-value cutoff}

\item{TERM2METAB}{2-columns data.frame of pathways and metabolites}

\item{TERM2NAME}{2-columns data.frame of pathways and pathway names}

\item{nPerm}{number of permutations}

\item{verbose}{verbose mode}

\item{gsea_by}{gsea type}

\item{seed}{set the seed}

\item{exponent}{exponent in the gsea approach}
}
\description{
Pathway analysis
Run enricher or gsea from clusterProfiler on pathways provided by NCBI Biosytems
}
