% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gen.R
\name{pca_gen}
\alias{pca_gen}
\title{PCA general
Scree plot, score and loading plot (PC1 vs PC2)}
\usage{
pca_gen(m_list, dirout, col_by = "class", scale = TRUE, include_QC = TRUE)
}
\arguments{
\item{m_list}{margheRita m_list}
}
\description{
PCA general
Scree plot, score and loading plot (PC1 vs PC2)
}
