% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_norm.R
\name{pca_norm}
\alias{pca_norm}
\title{PCA after normalizazion
Scree plot, score and loading plot (PC1 vs PC2)}
\usage{
pca_norm(df_norm, dirout)
}
\arguments{
\item{df_norm:}{dataframe after normalization}
}
\description{
PCA after normalizazion
Scree plot, score and loading plot (PC1 vs PC2)
}
