% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_postprocessing.R
\name{pca_postprocessing}
\alias{pca_postprocessing}
\title{PCA function post processing
Output score, loading plot (PC1 vs PC2) and scree plot}
\usage{
pca_postprocessing(df_post, dirout)
}
\arguments{
\item{df_post:}{dataframe of data was processed:imputation of missing value, pareto scaled}
}
\description{
PCA function post processing
Output score, loading plot (PC1 vs PC2) and scree plot
}
