% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_preprocessing.R
\name{pca_preprocessing}
\alias{pca_preprocessing}
\title{PCA function pre processing
Output score, loading plot (PC1 vs PC2) and scree plot}
\usage{
pca_preprocessing(df, dirout)
}
\arguments{
\item{df:}{dataframe of data in which the first three column are MS DIAL ID, RT and m/z, sample in colum and features (metabolites) in rows}
}
\description{
PCA function pre processing
Output score, loading plot (PC1 vs PC2) and scree plot
}
