% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_matching.R
\name{peak_matching}
\alias{peak_matching}
\title{MS/MS peaks matching}
\usage{
peak_matching(
  library_list = NULL,
  library_matched_features = NULL,
  RI_sample = NULL,
  ppm_err = 20,
  intensity = 20,
  RI_diff_type = c("rel", "abs")
)
}
\arguments{
\item{library_list}{an object structured as the element "lib_precursor" of the list returned by select_library()}

\item{library_matched_features}{a list of data frame, obtained by check_RT_mass() function. it is contains a list of metabolite with desired retention time and PPM error. each metabolite is a data frame of potential candidates with specific features; Feature_ID, RT_err, RT_flag, ppm_error, mass_status and mass_flag.}

\item{RI_sample}{it is generated by calc_RI() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{ppm_err}{A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{intensity}{maximum relative intensity}

\item{RI_diff_type}{type of relative intensity difference: absolute or relative}
}
\value{
A list with a data.frame of matched_peaks and matrices with scores of all-pairs comparison between peaks
}
\description{
MS/MS peaks matching
}
