% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_matching_onlyMSMS.R
\name{peak_matching_onlyMSMS}
\alias{peak_matching_onlyMSMS}
\title{Peak matching for metabolite families}
\usage{
peak_matching_onlyMSMS(
  library_list = NULL,
  RT_mass = NULL,
  RI_sample = NULL,
  ppm_err = 10,
  intensity = 30,
  RI_diff_type = c("rel", "abs")
)
}
\arguments{
\item{library_list}{an object structured as the element "lib_precursor" of the list returned by margherita_library()}

\item{RT_mass}{a list of data frame, obtained by check_RT_mass() function. it is contains a list of metabolite with desired retention time and PPM error. each metabolite is a data frame of potential candidates with specific features; Feature_ID, RT_err, RT_flag, ppm_error, mass_status and mass_flag.}

\item{RI_sample}{it is generated by RI_sample_data() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{ppm_err}{A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{intensity}{maximum relative intensity}

\item{RI_diff_type}{type of relative intensity difference: absolute or relative}
}
\description{
Peak matching for metabolite families
}
