% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2DPCA_fast.R
\name{Plot2DPCA_fast}
\alias{Plot2DPCA_fast}
\title{Create 2D PCA score and loading plots
Create 2D PCA score and loading plots choosing the principal components.}
\usage{
Plot2DPCA_fast(
  mRList,
  pcx = 1,
  pcy = 2,
  dirout = "./",
  col_by = "class",
  include_QC = TRUE
)
}
\arguments{
\item{mRList}{mRList object}

\item{pcx}{Specify the principal component on the x-axis}

\item{pcy}{Specify the principal component on the y-axis}

\item{dirout}{output directory}

\item{col_by}{(default class)}

\item{include_QC}{(default TRUE)}
}
\description{
Create 2D PCA score and loading plots
Create 2D PCA score and loading plots choosing the principal components.
}
