% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqn.R
\name{pqn}
\alias{pqn}
\title{PQN normalization}
\usage{
pqn(x, xref)
}
\arguments{
\item{x}{metabolomics profile}

\item{xref}{reference profile}
}
\value{
list of:
\itemize{
\item y = normalized profile;
\item se = size effects;
\item se_med = average size effect.
}
}
\description{
PQN normalization
}
\author{
Ettore Mosca
}
