% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_file.bkp.R
\name{read_input_file_bkp}
\alias{read_input_file_bkp}
\title{Read files (input and metadata) both excel and .csv, default .csv
import excel files input and metadata}
\usage{
read_input_file_bkp(
  input,
  metadata,
  split_QC = TRUE,
  mz_col = 3,
  rt_col = 2,
  data_start_col = 4
)
}
\description{
Read files (input and metadata) both excel and .csv, default .csv
import excel files input and metadata
}
