% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_lib_data.R
\name{RI_lib_data}
\alias{RI_lib_data}
\title{Calculate relative intensities and filter}
\usage{
RI_lib_data(reference_spectra, acceptable_RI = 10)
}
\arguments{
\item{reference_spectra}{list of data.frames with spectra}

\item{acceptable_RI}{maximum admitted relative intensity}
}
\description{
Calculate relative intensities and filter
}
