% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_sample_data.R
\name{RI_sample_data}
\alias{RI_sample_data}
\title{establishing new sample data by calculating relative intensity}
\usage{
RI_sample_data(feature_data_spectra)
}
\description{
establishing new sample data by calculating relative intensity
}
