% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_library.R
\name{select_library}
\alias{select_library}
\title{Library selection}
\usage{
select_library(
  column = c("HILIC", "LipC8", "pZIC", "RPLong", "RPShort"),
  mode = c("POS", "NEG"),
  accept_RI = 10
)
}
\description{
Library selection
}
