% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sample.R
\name{select_sample}
\alias{select_sample}
\title{sample selection}
\usage{
select_sample(mRList = NULL, n_column = c(1, 2, 3, 5), accept_RI = 10)
}
\arguments{
\item{mRList}{obtained by read_input_file() function}

\item{n_column}{it is selecting the required column for metabolite annotation which are; Feature_ID, rt, mz, MS_MS_spectrum.}

\item{accept_RI}{numeric parameter. the default value is 10. it is a maximum relative intensity that is kept in sample. since low intense peaks could be noise, it is filtering sample data set by deleting the relative intensity lower then accept_RI.}
}
\description{
sample selection
}
