% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sign_features.R
\name{select_sign_features}
\alias{select_sign_features}
\title{select significant features}
\usage{
select_sign_features(
  mRList = NULL,
  data.use = c("data", "data_ann"),
  test_method = "anova",
  test_value = "q",
  cutoff_value = 0.05
)
}
\arguments{
\item{mRList}{mRList object}

\item{data.use}{whether to use mRList$data (default) or mRList$data_ann, which is the annotated version obtained via metabolite_identification}

\item{test_method}{the name of the statistical test table contained in the mRlist object}

\item{test_value}{the column of the statistical test table to consider (e.g.: p or q)}

\item{cutoff_value}{the value of p-value cut-off to consider a feature as significant

#' @return a character vector with features names}
}
\description{
The output of this function can be passed in the features argument of the h_map function
}
