% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_effect.R
\name{size_effect}
\alias{size_effect}
\title{Size effect for pqn}
\usage{
size_effect(x = NULL, xref = NULL, robust = T)
}
\arguments{
\item{x}{metabolomics profile}

\item{xref}{reference profile}

\item{robust}{TRUE/FALSE}
}
\value{
list of:
\itemize{
\item se = size effects;
\item avg = average size effect.
}
}
\description{
Size effect for pqn
}
