% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitQC.R
\name{splitQC}
\alias{splitQC}
\title{Split dataframe and metadata in two: one for samples and one for QC}
\usage{
splitQC(m_list)
}
\arguments{
\item{m_list}{split metadata and dataframe for QC and for samples}
}
\value{

}
\description{
Split dataframe and metadata in two: one for samples and one for QC
}
