% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{Subset}
\usage{
subset(m_list)
}
\arguments{
\item{metadata}{split in metadata for QC and metadata for samples}

\item{calculate}{mean for each biological replicates}
}
\description{
Subset
}
