% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Univariate analysis}
\usage{
univariate(
  mRList = NULL,
  dirout = NULL,
  test_method = c("ttest", "Utest", "anova", "kruskal"),
  exp.levels = NULL,
  exp.factor = "class"
)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}

\item{test_method}{any between "ttest", "Utest", "anova" or "kruskal".}

\item{exp.levels}{the levels to be considered in the column specified by exp.factor}

\item{exp.factor}{column of mRList$sample_ann that defines the level for each sample}
}
\value{
mRList object with mRList$"testchosen" with univariate analysis
}
\description{
It applies T test, U test, Anova or Kruskal-Wallis test over dataset features.
}
\details{
In case of "anova", the results of stats::aov() will be analysed by stats::TukeyHSD()
}
