% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateEM.R
\name{univariateEM}
\alias{univariateEM}
\title{user has to Choose between T-test, U-test, Anova or Kruskal}
\usage{
univariateEM(
  mRList,
  dirout = "./",
  test_method = c("ttest", "anova", "Utest", "kruskal"),
  paired = c("FALSE", "TRUE"),
  group_factor = "class"
)
}
\arguments{
\item{mRList}{(dataframe of samples only biological replicates, technical replicates were collapsed before)}
}
\description{
user has to Choose between T-test, U-test, Anova or Kruskal
}
