% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_1spectra.R
\name{visualize_1spectra}
\alias{visualize_1spectra}
\title{Visualize the spectra of good candidates}
\usage{
visualize_1spectra(
  feature_id = NULL,
  peak_id = NULL,
  RI_sample = NULL,
  RI_lib = NULL,
  lib_peaks_data = NULL,
  out_dir = "./"
)
}
\arguments{
\item{feature_id}{feature id}

\item{peak_id}{peak id}

\item{RI_sample}{it is generated by RI_sample_data() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{RI_lib}{it is generated by RI_lib_data() or select_library() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks. note that in house library of the margheRita is already stored by relative intensity and filtered by select_library() function so, in this case RI_lib is lib_peaks which is generated by select_library() function.}

\item{lib_peaks_data}{it is generated by select_library() function. it is providing information regarding the peaks such as ID, CAS number, Name and POS/NEG collision energy.}

\item{out_dir}{output directory}
}
\description{
Visualize the spectra of good candidates
}
