% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_associated_spectra.R
\name{visualize_associated_spectra}
\alias{visualize_associated_spectra}
\title{MS/MS Spectra Visualization
Visualize the MS/MS spectra of the features associated with a metabolite.}
\usage{
visualize_associated_spectra(
  mRList = NULL,
  mR_library = NULL,
  metabolite_id = NULL,
  dirout = NULL,
  type = c("mirrored", "overlapped"),
  peak_number = FALSE
)
}
\arguments{
\item{mRList}{mRList object that contains the results of metabolite identification}

\item{mR_library}{library list used during annotation}

\item{metabolite_id}{ID of the metabolite to be visualized}

\item{dirout}{output directory}

\item{type}{mirrored or overlapped visualization}

\item{peak_number}{if TRUE write the peak number}
}
\description{
MS/MS Spectra Visualization
Visualize the MS/MS spectra of the features associated with a metabolite.
}
