% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_spectra.R
\name{visualize_spectra}
\alias{visualize_spectra}
\title{Visualize the spectra of good candidates}
\usage{
visualize_spectra(annotation_list, x = NULL, out_dir = "./")
}
\arguments{
\item{annotation_list}{annotation_list}

\item{x}{one or more metabolites to visualize}

\item{out_dir}{output directory}
}
\description{
Visualize the spectra of good candidates
}
