% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_spectra2.R
\name{visualize_spectra2}
\alias{visualize_spectra2}
\title{Visualize the spectra of good candidates}
\usage{
visualize_spectra2(x = NULL, RI_sample = NULL, RI_lib = NULL)
}
\arguments{
\item{x}{candidates}

\item{RI_sample}{it is generated by RI_sample_data() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{RI_lib}{it is generated by RI_lib_data() or select_library() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks. note that in house library of the margheRita is already stored by relative intensity and filtered by select_library() function so, in this case RI_lib is lib_peaks which is generated by select_library() function.}
}
\description{
Visualize the spectra of good candidates
}
