% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.PomaSummarizedExperiment.R
\name{as.PomaSummarizedExperiment}
\alias{as.PomaSummarizedExperiment}
\title{Transforms mRList into PomaSummarizedExperiment}
\usage{
as.PomaSummarizedExperiment(mRList = NULL)
}
\arguments{
\item{mRList}{mRList object}
}
\value{
PomaSummarizedExperiment
}
\description{
Transforms mRList into PomaSummarizedExperiment
}
