% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_reference.R
\name{calc_reference}
\alias{calc_reference}
\title{calculate reference}
\usage{
calc_reference(
  m_list,
  sample_col = "class",
  sample_class = "QC",
  approach = c("median", "mean")
)
}
\description{
calculate reference
}
