% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_intense_peak.R
\name{check_intense_peak}
\alias{check_intense_peak}
\title{check the ppm error for the intense peaks}
\usage{
check_intense_peak(
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  reference = NULL,
  lib_peaks_data = NULL,
  n_peaks = 1,
  acceptable_PPM_err = 10,
  mode = c("POS", "NEG")
)
}
\arguments{
\item{RT_mass}{RT_mass}

\item{RI_lib}{RI_lib}

\item{RI_sample}{RI_sample}

\item{reference}{reference}

\item{lib_peaks_data}{lib_peaks_data}

\item{n_peaks}{n_peaks}

\item{acceptable_PPM_err}{acceptable_PPM_err}

\item{mode}{mode}
}
\description{
check the ppm error for the intense peaks
}
