% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mass.R
\name{check_mass}
\alias{check_mass}
\title{check_mass}
\usage{
check_mass(
  feature_data = NULL,
  reference = NULL,
  unaccept_flag = 15,
  accept_flag = 5,
  suffer_flag = 10
)
}
\description{
check_mass
}
