% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RT_mass.R
\name{check_RT_mass}
\alias{check_RT_mass}
\title{Merge candidates with appropriate retention time and PPM error}
\usage{
check_RT_mass(RT = NULL, mass = NULL)
}
\arguments{
\item{RT}{A list of library ID each contain a data frame of sample ID with ideal retention time. it is obtained from previous function.}

\item{mass}{A list of library ID each contain a data frame of sample ID with ideal PPM error. it is obtained from previous function.}
}
\value{
data.frame with candidates
}
\description{
Merge candidates with appropriate retention time and PPM error
}
