% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RT_mass.R
\name{check_RT_mass}
\alias{check_RT_mass}
\title{Merging ideal candidate in term of Retention Time and PPM error}
\usage{
check_RT_mass(RT = NULL, mass = NULL, reference = NULL)
}
\description{
Merging ideal candidate in term of Retention Time and PPM error
}
