% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse
Subsetting metadata and data creating QC and samples data and metadata
Collapse technical replicates for samples.}
\usage{
collapse(m_list)
}
\arguments{
\item{For}{each sample mean was calculated to collapse technical replicates}
}
\description{
Collapse
Subsetting metadata and data creating QC and samples data and metadata
Collapse technical replicates for samples.
}
