% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_tech_rep.R
\name{collapse_tech_rep}
\alias{collapse_tech_rep}
\title{Collapse technical replicates for samples.}
\usage{
collapse_tech_rep(m_list, remove.QC = FALSE)
}
\arguments{
\item{For}{each sample mean was calculated to collapse technical replicates}
}
\description{
Collapse technical replicates for samples.
}
