% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_tech_rep.R
\name{collapse_tech_rep}
\alias{collapse_tech_rep}
\title{Collapse technical replicates of each biological replicate of samples, calculating the mean.}
\usage{
collapse_tech_rep(m_list, remove.QC = FALSE)
}
\arguments{
\item{m_list}{mRList}
}
\value{
m_list with only biological replicates
}
\description{
Collapse technical replicates of each biological replicate of samples, calculating the mean.
}
