% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Calculate of coefficient of variation for each metabolite}
\usage{
CV(mRList, dirout = "./")
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}
}
\value{
filtered mRList object
}
\description{
Calculate the coefficient of variance (CV) for samples and QC (quality control) to assess the reliability of the experiment, as:
CV = (SD - mean) * 100.
Only metabolites with sample CV < QC CV are kept.
}
\examples{
##library(dataset.margheRita)
##dataset(norm_pos)
mRList<-CV(mRList)

}
