% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Select QC_samples and samples.
"QC" and "m" that indicates quality control samples and biological replicates in the title were used to select, respectively}
\usage{
CV(m_list, dirout)
}
\arguments{
\item{df_post}{split into df_QC (dataframe of quality control) and df_samples (df of samples)}
}
\description{
Select QC_samples and samples.
"QC" and "m" that indicates quality control samples and biological replicates in the title were used to select, respectively
}
