% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_map.R
\name{h_map}
\alias{h_map}
\title{Draw an heatmap}
\usage{
h_map(
  mRList = NULL,
  column_ann = "class",
  data.use = c("data", "data_ann"),
  col_ann = NULL,
  col = NULL,
  scale_features = TRUE,
  features = NULL,
  samples = NULL,
  top = 200,
  use_top_annotated_metab = FALSE,
  test_method = "anova",
  test_value = "q",
  truncate_row_labels = FALSE,
  rm_outliers = FALSE,
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{column_ann}{sample_ann column for sample annotation}

\item{data.use}{whether to use mRList$data (default) or mRList$data_ann, which is the annotated version obtained via metabolite_identification}

\item{col_ann}{set of colors for sample annotation}

\item{col}{set of color for data values}

\item{scale_features}{whether to scale features or not}

\item{features}{names of features to plot (optional)}

\item{samples}{samples to consider (optional)}

\item{top}{only the top most variable features are plotted (if features is NULL)}

\item{use_top_annotated_metab}{if TRUE, it'll use the best name for the metabolite from the annotation, considering the level and also the top significant specified in the following arguments.}

\item{test_method}{the name of the statistical test table contained in the mRlist object}

\item{test_value}{the column of the statistical test table to consider (e.g.: p or q)}

\item{truncate_row_labels}{FALSE or a integer that specifies the number of characters to keep}

\item{rm_outliers}{remove outliers when setting colours}

\item{...}{further arguments for ComplexHeatmap::Heatmap}
}
\description{
Draw an heatmap
}
