% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_map.R
\name{h_map}
\alias{h_map}
\title{Draw an heatmap}
\usage{
h_map(
  mRList = NULL,
  dirout = NULL,
  column_ann = "class",
  col_ann = NULL,
  col = NULL,
  scale_features = TRUE,
  features = NULL,
  samples = NULL,
  top = 20,
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}

\item{column_ann}{sample_ann column for sample annotation}

\item{col_ann}{set of colors for sample annotation}

\item{col}{set of color for data values}

\item{scale_features}{whether to scale features or not}

\item{features}{names of features to plot (optiooal)}

\item{samples}{samples to consider (optional)}

\item{top}{only the top most variable features are plotted (if features is NULL)}

\item{...}{further arguments for ComplexHeatmap::Heatmap}
}
\description{
Draw an heatmap
}
