% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatscatter_chromatography.R
\name{heatscatter_chromatography}
\alias{heatscatter_chromatography}
\title{Draw the "heatscatter chromatography"}
\usage{
heatscatter_chromatography(
  mRList = NULL,
  mz_limits = NULL,
  rt_limits = NULL,
  sample = NULL,
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{mz_limits}{numeric vector with minimum and maximum rt values}

\item{rt_limits}{numeric vector with minimum and maximum rt values}

\item{sample}{if not null, consider only the given samples}

\item{...}{further arguments to LSD::heatscatter}
}
\description{
Draw the "heatscatter chromatography"
}
