% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{missing values imputation}
\usage{
imputation(m_list, seed = NULL, a = 0.1, b = 0.25, n = 100)
}
\arguments{
\item{dataframe}{of input contains intensity of all features according to samples}
}
\description{
missing values imputation
}
