% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Missing values imputation}
\usage{
imputation(mRList = NULL, seed = NULL, a = 0.1, b = 0.25)
}
\arguments{
\item{mRList}{mRList object}

\item{seed}{set this seed with set.seed}

\item{a}{minimum increase factor}

\item{b}{maxmimum increase factor}
}
\value{
mRList object with imputed values
}
\description{
the NA values are replaced with a fraction of the minimum value of that row (min_r), sampled in the interval \eqn{[min_r * a, min_r * b]}
}
