% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_z_filtering.R
\name{m_z_filtering}
\alias{m_z_filtering}
\title{Filter m/z of metabolites}
\usage{
m_z_filtering(
  mRList,
  do_plot = TRUE,
  out_file = "mz_quality.png",
  lower_quality_mass_acc = 0.4,
  upper_quality_mass_acc = 0.8,
  color = "black",
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{do_plot}{whether to plot or not}

\item{out_file}{output file name}

\item{lower_quality_mass_acc}{lower quality value}

\item{upper_quality_mass_acc}{upper quality value}

\item{color}{point color}

\item{...}{further arguments to plot function}
}
\value{
filtered mRList
}
\description{
Filter m/z of metabolites
}
