% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median_stdev_samples.R
\name{mean_median_stdev_samples}
\alias{mean_median_stdev_samples}
\title{Calculate mean, median and standard deviation for each biological group.}
\usage{
mean_median_stdev_samples(mRList = NULL, dirout = NULL)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}
}
\description{
Calculate mean, median and standard deviation for each biological group under study. This function works after collapse_tech_rep.
}
