% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median_stdev_samples.R
\name{mean_median_stdev_samples}
\alias{mean_median_stdev_samples}
\title{Calculate mean and median for each biological group.}
\usage{
mean_median_stdev_samples(mRList, dirout = "./", write_output = FALSE)
}
\arguments{
\item{mRList}{mRList contains only biological replicates}
}
\description{
Calculate mean, median and standard deviation for each biological group under study.

This function works after collapse_tech_rep.

write_output (default FALSE), if TRUE table, as .csv format, is written and saved.
}
