% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_stats_with_features.R
\name{merge_stats_with_features}
\alias{merge_stats_with_features}
\title{Get significant features}
\usage{
merge_stats_with_features(mRList = NULL, feature_stats = NULL)
}
\arguments{
\item{mRList}{mRList object}

\item{feature_stats}{name of an element of mRList object containing results of statistical test or a custom data frame with Feature_ID as row names.}
}
\value{
a data.frame
}
\description{
Merge mRList$metab_ann with feature_stats
}
