% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_boxplot.R
\name{metab_boxplot}
\alias{metab_boxplot}
\title{metab_boxplot}
\usage{
metab_boxplot(
  mRList = NULL,
  dirout = "./",
  features = NULL,
  col_by = "class",
  group = "class"
)
}
\arguments{
\item{mRList}{mRList}
}
\description{
metab_boxplot
}
