% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_boxplot.R
\name{metab_boxplot}
\alias{metab_boxplot}
\title{Boxplot of metabolites}
\usage{
metab_boxplot(
  mRList = NULL,
  dirout = "./",
  features = NULL,
  col_by = "class",
  group = "class"
)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}

\item{features}{vector with a list of metabolites to graph}

\item{col_by}{define how to color the boxplot, default=class}

\item{group}{define what you want to compare}
}
\value{
.png of boxplot
}
\description{
Plot boxplot of metabolites, and save them as .png.
}
\examples{
##library(dataset.margheRita)
##dataset(norm_pos)
metab_boxplot(mRList,features=mRLIst$data[1:10,],col_by="class",group="class")
}
