% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_boxplot.R
\name{metab_boxplot}
\alias{metab_boxplot}
\title{metab_boxplot}
\usage{
metab_boxplot(
  m_list,
  dirout = "./",
  features = NULL,
  col_by = "class",
  group = "class"
)
}
\arguments{
\item{m_list}{m_list}
}
\description{
metab_boxplot
}
