% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_annotation.R
\docType{data}
\name{metabolite_annotation}
\alias{metabolite_annotation}
\title{metabolite_annotation
map between metabolite name and various stantard identifiers}
\format{
An object of class \code{data.frame} with 834 rows and 15 columns.
}
\usage{
metabolite_annotation
}
\description{
metabolite_annotation
map between metabolite name and various stantard identifiers
}
\keyword{datasets}
