% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_annotation.R
\name{metabolite_annotation}
\alias{metabolite_annotation}
\title{Metabolite Annotation}
\usage{
metabolite_annotation(
  mRList = NULL,
  library_list = NULL,
  feature_spectra = NULL,
  rt_err_thr = 1,
  unaccept_flag = 15,
  accept_flag = 5,
  suffer_flag = 10,
  acceptable_RI = 10,
  n_peaks = 1,
  acceptable_PPM_err = 10,
  mode = NULL
)
}
\description{
Metabolite Annotation
}
