% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_annotation 2.R,
%   R/metabolite_annotation.R
\name{metabolite_annotation}
\alias{metabolite_annotation}
\title{metabolite_annotation
map between metabolite name and various stantard identifiers}
\usage{
metabolite_annotation(
  mR_list = NULL,
  reference = NULL,
  feature_spectra = NULL,
  reference_spectra = NULL,
  rt_err_thr = 1,
  unaccept_flag = 15,
  accept_flag = 5,
  suffer_flag = 10,
  acceptable_RI = 10,
  n_peaks = 1,
  acceptable_PPM_err = 10
)

metabolite_annotation(
  mR_list = NULL,
  reference = NULL,
  feature_spectra = NULL,
  reference_spectra = NULL,
  rt_err_thr = 1,
  unaccept_flag = 15,
  accept_flag = 5,
  suffer_flag = 10,
  acceptable_RI = 10,
  n_peaks = 1,
  acceptable_PPM_err = 10
)
}
\arguments{
\item{reference}{data.frame with columns "rt", "mz", "Name", "MS_MS_spectrum"}

\item{mRList}{mRList object}
}
\value{
the annotation list, which has three elements... 1) named list of data.frames; each element of the list corresponds to an element of reference and each data.frame containes the candidates
}
\description{
metabolite_annotation
map between metabolite name and various stantard identifiers

Metabolite Annotation
}
