% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_family.R
\name{metabolite_family}
\alias{metabolite_family}
\title{Metabolite Family}
\usage{
metabolite_family(
  feature_data = NULL,
  reference = NULL,
  rt_err_thr = 2,
  unaccept_flag = 1000,
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  lib_peaks_data = NULL,
  mode = c("POS", "NEG")
)
}
\description{
Metabolite Family
}
