% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_family.R
\name{metabolite_family}
\alias{metabolite_family}
\title{Find metabolite families}
\usage{
metabolite_family(
  reference = NULL,
  feature_data = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  lib_peaks_data = NULL,
  rt_err_thr = 1,
  unaccept_flag = 10,
  mode = c("NEG", "POS"),
  ppm_err = 10,
  intensity = 30
)
}
\arguments{
\item{reference}{an object structured as the element "lib_precursor" of the list returned by margherita_library()}

\item{feature_data}{an object structured as the elment "metab_ann" of the mRList object}

\item{RI_lib}{it is generated by RI_lib_data() or select_library() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks. note that in house library of the margheRita is already stored by relative intensity and filtered by select_library() function so, in this case RI_lib is lib_peaks which is generated by select_library() function.}

\item{RI_sample}{it is generated by RI_sample_data() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{lib_peaks_data}{it is generated by select_library() function. it is providing information regarding the peaks such as ID, CAS number, Name and POS/NEG collision energy.}

\item{rt_err_thr}{threshold over the retention time error}

\item{unaccept_flag}{A number with default value of 15. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{mode}{mode could be set in positive or negative state. positive mode select positive collision energy and mz in positive mode.}

\item{ppm_err}{A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{intensity}{maximum relative intensity}
}
\description{
Find metabolite families
}
