% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_identification.R
\name{metabolite_identification}
\alias{metabolite_identification}
\title{Metabolite Identification
Run the metabolite identification pipeline}
\usage{
metabolite_identification(
  mRList = NULL,
  library_list = NULL,
  rt_err = 1,
  unaccept_flag = 20,
  accept_flag = 5,
  suffer_flag = 10,
  min_RI = 10,
  ppm_err = 20,
  RI_err = 20,
  RI_err_type = "rel",
  filter = TRUE
)
}
\arguments{
\item{mRList}{mRList object}

\item{library_list}{library data obtained from margherita_library() function}

\item{rt_err}{threshold over the retention time error}

\item{unaccept_flag}{A number with default value of 15. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{accept_flag}{A number with default value of 5. PPM errors < accept_flag will be tagged as "super", while those > accept_flag and < suffer_flag will be tagged as "acceptable"}

\item{suffer_flag}{A number with default value of 10. PM errors above this value and < unaccept_flag will be tagged as "suffer"}

\item{min_RI}{numeric parameter. the default value is 10. it is a maximum relative intensity that is kept in sample. since low intense
peaks could be noise, it is filtering sample dataset by deleting the relative intensity lower then accept_RI.}

\item{ppm_err}{A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{RI_err}{maximum absolute RI difference between MS/MS peaks of sample and library}

\item{RI_err_type}{type of RI error calculation.}

\item{filter}{whether to filter metabolite-feature associations or not.}
}
\value{
data.frame with matches between features and library metabolites
}
\description{
Metabolite Identification
Run the metabolite identification pipeline
}
