% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mR_pca.R
\name{mR_pca}
\alias{mR_pca}
\title{PCA}
\usage{
mR_pca(
  mRList,
  dirout,
  col_by = "class",
  method = "svd",
  scaling = c("none", "pareto", "vector", "uv"),
  center = TRUE,
  include_QC = TRUE,
  top = Inf,
  nPcs,
  write_output = FALSE,
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}

\item{col_by}{(default class)}

\item{method}{see pcaMethods::pca()}

\item{scaling}{see pcaMethods::prep()}

\item{center}{whether to center the data or not}

\item{include_QC}{(default TRUE)}

\item{top}{only the top most varying features will be used}

\item{nPcs}{number of principal components}

\item{write_output}{if score and loading table have to be saved (default=FALSE)}

\item{...}{further argments to pcaMethods::pca}
}
\value{
mRList object with "pca" element. Images of screeplot and pairs, table of score,loading and variance can be saved if write_output=TRUE.
}
\description{
This function allows to perform PCA.
}
\examples{
##library(dataset.margheRita)
##dataset(norm_pos)
pca_fast(mRList, col_by="class", scaling="pareto",nPcs=5,write_output=FALSE)
}
