% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_check.R
\name{norm_check}
\alias{norm_check}
\title{Shapiro-Wilk's test is used for testing the normal distribution of metabolites lo faccio sul subsetting Samples}
\usage{
norm_check(m_list)
}
\arguments{
\item{m_list}{margheRita m_list}
}
\description{
Shapiro-Wilk's test is used for testing the normal distribution of metabolites lo faccio sul subsetting Samples
}
