% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_profiles.R
\name{normalize_profiles}
\alias{normalize_profiles}
\title{Normalize profiles}
\usage{
normalize_profiles(m_list, method = c("log", "reference", "pqn"))
}
\arguments{
\item{method}{"log": log2(x+1); "reference": divide each sample by its reference; "pqn": probailistic quotient normalization;}

\item{m_list$data}{features-by-samples matrix of metabolite levels;}

\item{reference}{reference profile;}
}
\description{
Normalize profiles
}
