% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_profiles.R
\name{normalize_profiles}
\alias{normalize_profiles}
\title{Normalize profiles}
\usage{
normalize_profiles(X, method = c("log", "reference", "pqn"), reference = NULL)
}
\arguments{
\item{X}{features-by-samples matrix of metabolite levels}

\item{method}{log log-normalization; reference: divide each sample by its reference; pqn probailistic quotient normalization #' @param reference reference profile}
}
\description{
Normalize profiles
}
