% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{pareto scaling of the dataframe after missing value correction}
\usage{
pareto(df)
}
\arguments{
\item{df}{data.frame without the first three columns (MS DIAL, RT and m/z)}
}
\description{
pareto scaling of the dataframe after missing value correction
}
