% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Pareto scaling
Perform data scaling with Pareto approach.}
\usage{
pareto(X, centering = TRUE)
}
\arguments{
\item{X}{is dataframe that contains only data, without the columns that contain ID metabolite, m/z, RT, MS and MS/MS information}

\item{centering}{whether to center the data or not. Defaults to TRUE.}
}
\description{
Pareto scaling
Perform data scaling with Pareto approach.
}
