% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{Pathway analysis
Run clusterProfiler on pathways provided by NCBI Biosytems to perform ORA or MSEA}
\usage{
pathway_analysis(
  in_list = NULL,
  universe = NULL,
  type = c("ora", "msea"),
  tax_id = 9606,
  include_general_pathways = FALSE,
  pvalueCutoff = 1,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 1,
  TERM2METAB = NULL,
  TERM2NAME = NULL,
  nPerm = 1000,
  verbose = TRUE,
  msea_by = "fgsea",
  seed = FALSE,
  exponent = 1
)
}
\arguments{
\item{in_list}{a vector of metabolite id for ORA or a ranked named vector with scores for MSEA, where names are metabolite ids}

\item{universe}{the set of all metabolites, required by ORA}

\item{type}{"ora" or "msea"}

\item{tax_id}{tax id}

\item{include_general_pathways}{whether to include general pathways or not}

\item{pvalueCutoff}{p-value cutoff}

\item{pAdjustMethod}{p-value adjustment method}

\item{minGSSize}{minimum metabolite set size}

\item{maxGSSize}{maximum metabolite set size}

\item{qvalueCutoff}{q-value cutoff}

\item{TERM2METAB}{2-columns data.frame of pathways and metabolites}

\item{TERM2NAME}{2-columns data.frame of pathways and pathway names}

\item{nPerm}{number of permutations}

\item{verbose}{verbose mode}

\item{msea_by}{msea type}

\item{seed}{set the seed}

\item{exponent}{exponent in the msea approach}
}
\value{
A list with two elements
\itemize{
\item res: gseaResult object or an enrichResult instance
\item term_annotation: data.frame with pathway annotation
}
}
\description{
Pathway analysis
Run clusterProfiler on pathways provided by NCBI Biosytems to perform ORA or MSEA
}
