% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{pathway analysis
Run enricher or gsea from cluster profiler on BIOCYC data provided by
NCBI Biosytems}
\usage{
pathway_analysis(
  in_list = NULL,
  universe = NULL,
  type = c("ora", "gsea"),
  tax_id = 9606,
  include_general_pathways = FALSE,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  TERM2METAB = NULL,
  TERM2NAME = NULL,
  nPerm = 1000,
  verbose = TRUE,
  gsea_by = "fgsea",
  seed = FALSE,
  exponent = 1
)
}
\description{
pathway analysis
Run enricher or gsea from cluster profiler on BIOCYC data provided by
NCBI Biosytems
}
