% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_fast.R
\name{pca_fast}
\alias{pca_fast}
\title{PCA
Function that allows to perform PCA. Graphs of Scree plot and pairs of the first 10 components were plotted. Score and loading plot, choosing pcx and pcy, were plotted launching function plot2DPCA after pca function.}
\usage{
pca_fast(
  mRList,
  dirout,
  col_by = "class",
  method = "svd",
  scaling = c("none", "pareto", "vector", "uv"),
  center = TRUE,
  include_QC = TRUE,
  top = Inf,
  nPcs = 5,
  ...
)
}
\arguments{
\item{mRList}{mRList object}

\item{dirout}{output directory}

\item{col_by}{(default class)}

\item{method}{see pcaMethods::pca()}

\item{scaling}{see pcaMethods::prep()}

\item{center}{whether to center the data or not}

\item{include_QC}{(default TRUE)}

\item{top}{only the top most varying features will be used}

\item{nPcs}{number of principal components}

\item{...}{further argments to pcaMethods::pca}
}
\value{
mRList object with "pca" element
}
\description{
PCA
Function that allows to perform PCA. Graphs of Scree plot and pairs of the first 10 components were plotted. Score and loading plot, choosing pcx and pcy, were plotted launching function plot2DPCA after pca function.
}
