% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gen.R
\name{pca_gen}
\alias{pca_gen}
\title{PCA analysis}
\usage{
pca_gen(
  mRList,
  dirout,
  col_by = "class",
  scaling = c("none", "Pareto", "uv"),
  include_QC = TRUE,
  type = c("component", "coordinate"),
  dist.method = "euclidean",
  top = Inf,
  write_output = FALSE
)
}
\arguments{
\item{mRList}{margheRita mRList}

\item{col_by}{(default class)}

\item{scaling}{choose the scaling method "none","pareto" or "uv"}

\item{include_QC}{(default TRUE)}

\item{write_output}{(default =FALSE) if it turns on TRUE tables as .csv of score and loading will be saved}
}
\value{
Graphs of screeplot and pairs; table as .csv if write_output turns TRUE

scoreplot and loading plot after plot2DPCA function has to be launch. User has to choose the PCs to plot graphs
}
\description{
Function that allows to perform PCA. Graphs of Scree plot and pairs of the first 10 components were plotted

Score and loading plot, choosing pcx and pcy, were plotted launching function plot2DPCA after pca function
}
