% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_matching.R
\name{peak_matching}
\alias{peak_matching}
\title{peak matching}
\usage{
peak_matching(
  feature_data = NULL,
  reference = NULL,
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  lib_peaks_data = NULL,
  mode = c("POS", "NEG"),
  ppm_err = 10,
  intensity = 30
)
}
\description{
peak matching
}
