% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_matching.R
\name{peak_matching}
\alias{peak_matching}
\title{Peak matching for metabolite families}
\usage{
peak_matching(
  reference = NULL,
  RT_mass = NULL,
  RI_lib = NULL,
  RI_sample = NULL,
  lib_peaks_data = NULL,
  mode = c("POS", "NEG"),
  ppm_err = 10,
  intensity = 30,
  RI_diff_type = c("rel", "abs")
)
}
\arguments{
\item{reference}{an object structured as the element "lib_precursor" of the list returned by margherita_library()}

\item{RT_mass}{a list of data frame, obtained by check_RT_mass() function. it is contains a list of metabolite with desired retention time and PPM error. each metabolite is a data frame of potential candidates with specific features; Feature_ID, RT_err, RT_flag, ppm_error, mass_status and mass_flag.}

\item{RI_lib}{it is generated by RI_lib_data() or select_library() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks. note that in house library of the margheRita is already stored by relative intensity and filtered by select_library() function so, in this case RI_lib is lib_peaks which is generated by select_library() function.}

\item{RI_sample}{it is generated by RI_sample_data() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.}

\item{lib_peaks_data}{it is generated by select_library() function. it is providing information regarding the peaks such as ID, CAS number, Name and POS/NEG collision energy.}

\item{mode}{mode could be set in positive or negative state. positive mode select positive collision energy and mz in positive mode.}

\item{ppm_err}{A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.}

\item{intensity}{maximum relative intensity}

\item{RI_diff_type}{type of relative intensity difference: absolute or relative}
}
\description{
Peak matching for metabolite families
}
