% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gen.R
\name{Plot2DPCA}
\alias{Plot2DPCA}
\title{Create 2D PCA score and loading plots}
\usage{
Plot2DPCA(mRList, pcx, pcy, dirout, col_by = "class", include_QC = T)
}
\arguments{
\item{mRList}{Input name of the created object}

\item{pcx}{Specify the principal component on the x-axis}

\item{pcy}{Specify the principal component on the y-axis}

\item{scoreplot}{input a name for the score plot}

\item{loadingplot}{input for the loading plot}

\item{format}{Select the image format, "png".}
}
\description{
Create 2D PCA score and loading plots choosing the principal components.
}
