% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_file.R
\name{read_input_file}
\alias{read_input_file}
\title{Create an mRList from feature data and sample information}
\usage{
read_input_file(
  feature_file = NULL,
  sample_file = NULL,
  MSDial = TRUE,
  split_QC = TRUE,
  mz_col = NULL,
  rt_col = NULL,
  data_start_col = NULL,
  MS_MS_column = NULL,
  ...
)
}
\arguments{
\item{feature_file}{tab-delimited text files with feature abundances, mz, rt and MS/MS spectra}

\item{sample_file}{sample information file}

\item{MSDial}{(default=TRUE) the input file is MS-Dial native format}

\item{split_QC}{(default=TRUE) split input and metadata in two one for QC and the other for data}

\item{mz_col}{column that contains m/z value}

\item{rt_col}{column that contains rt value}

\item{data_start_col}{column from which data starts}

\item{MS_MS_column}{column that contains the MS/MS spectrum}

\item{...}{further arguments to read.delim}
}
\value{
mRList object that contain data and metadata.
}
\description{
Import data and metadata as excel or .csv or .txt format
}
