% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_file.R
\name{read_input_file}
\alias{read_input_file}
\title{Import data and metadata}
\usage{
read_input_file(
  input,
  metadata,
  split_QC = TRUE,
  mz_col = 3,
  rt_col = 2,
  data_start_col = 4,
  MS_MS_column = NULL,
  type = c("excel", "txt"),
  ...
)
}
\arguments{
\item{input}{data file name}

\item{metadata}{metadata file name}

\item{split_QC}{(default=TRUE) split input and metadata in two one for QC and the other for data}

\item{mz_col}{column that contains m/z value}

\item{rt_col}{column that contains rt value}

\item{data_start_col}{column from which data starts}

\item{MS_MS_column}{column that contains the MS/MS spectrum}

\item{type}{type of imported files excel or .csv or .txt}

\item{...}{further arguments to read.table}
}
\description{
Import data and metadata as excel or .csv or .txt format
}
