% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_file.R
\name{read_input_file}
\alias{read_input_file}
\title{Read files (input and metadata) both excel and .csv, default .csv
import excel files input and metadata}
\usage{
read_input_file(input, metadata, split_QC = TRUE)
}
\description{
Read files (input and metadata) both excel and .csv, default .csv
import excel files input and metadata
}
