% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_samples.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples from mRList}
\usage{
remove_samples(mRList = NULL, ids = NULL, column = NULL)
}
\arguments{
\item{mRList}{object}

\item{ids}{rownames of $sample_ann to be removed. Otherwise, if "column" is given, all samples corresponding to the values appearing in the specfied column of $sample_ann will be removed.}

\item{column}{a column of $sample_ann to be used for selecting samples to be removed.}
}
\description{
Remove samples from mRList
}
