% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_lib_data.R
\name{RI_lib_data}
\alias{RI_lib_data}
\title{establishing new library data by calculating relative intensity}
\usage{
RI_lib_data(reference_spectra, RT_mass)
}
\description{
establishing new library data by calculating relative intensity
}
