% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_sample_data.R
\name{RI_sample_data}
\alias{RI_sample_data}
\title{converting intensity to relative intensity}
\usage{
RI_sample_data(feature_spectra, accept_RI = 10)
}
\arguments{
\item{feature_spectra}{it is a list of sample ID which each contains list of its m/z and Intensity.}

\item{accept_RI}{numeric parameter. the default value is 10. it is a maximum relative intensity that is kept in sample. since low intense peaks could be noise, it is filtering sample dataset by deleting the relative intensity lower then accept_RI.}
}
\description{
converting intensity to relative intensity
}
