% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLA.R
\name{RLA}
\alias{RLA}
\title{Relative Log Abudance}
\usage{
RLA(x, robust = TRUE)
}
\arguments{
\item{x}{metabolites-by-samples matrix of metabolite levels}
}
\description{
Relative Log Abudance
}
\author{
Ettore Mosca (CNR-ITB)
}
