% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLA.R
\name{RLA}
\alias{RLA}
\title{Relative Log Abudance}
\usage{
RLA(
  mRList,
  include_QC = FALSE,
  logged = FALSE,
  robust = TRUE,
  do_plot = FALSE,
  out_dir = "./",
  ...
)
}
\arguments{
\item{mRList}{margheRita list}
}
\description{
Relative Log Abudance
}
\author{
Ettore Mosca (CNR-ITB)
}
