% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_library.R
\name{select_library}
\alias{select_library}
\title{Library selection}
\usage{
select_library(
  source = c("margheRita", "MS-DIAL"),
  column = NULL,
  mode = c("POS", "NEG"),
  accept_RI = 10
)
}
\arguments{
\item{source}{"margheRita" or "MS-DIAL"}

\item{column}{column type, only if source=="margheRita". Possible values are: HILIC, LipC8, pZIC, RPLong, RPShort. Based on the type of the column, the retention time of same metabolite could be different.}

\item{mode}{mode could be set in positive or negative state. positive mode select positive collision energy and mz in positive mode.}

\item{accept_RI}{numeric parameter. the default value is 10. it is a maximum relative intensity that we keep in library. since low intense peaks could be noise, it is filtering the library by deleting the relative intensity lower then accept_RI.}
}
\value{
A list with:
\itemize{
\item precursors library: it is a list of metabolites which each contains information of retention time and mz with their specific IDs, CAS numbers and names.
\item lib_peaks_data:it is a list of metabolites which each contains information of collision energy with their specific IDs, CAS numbers and names.
\item lib_peaks: it is a list of metabolites with same IDs as lib_peaks_data which each contains a list of a peaks with mz and relative intensity.
\item key_field: the key field of the library
}
}
\description{
Library selection
}
