% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sign_features.R
\name{select_sign_features}
\alias{select_sign_features}
\title{select significant features}
\usage{
select_sign_features(
  mRList = NULL,
  test_method = "anova",
  test_value = "q",
  cutoff_value = 0.05,
  feature_id = "Feature_ID",
  values = FALSE,
  split_by_semicolon = FALSE
)
}
\arguments{
\item{mRList}{mRList object}

\item{test_method}{the name of the statistical test table contained in the mRlist object}

\item{test_value}{the column of the statistical test table to consider (e.g.: p or q)}

\item{cutoff_value}{the value of p-value cut-off to consider a feature as significant}

\item{feature_id}{"Feature_ID" or "Name" from the mRList$data_ann}

\item{values}{if TRUE a named vector with test_value values will be given in output; the beast value will be considered in case of 1-to-n mappings}

\item{split_by_semicolon}{if TRUE, the names will be split by semicolons.}
}
\value{
a character vector with features names
}
\description{
The output of this function can be passed in the features argument of the h_map function
}
