% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_effect.R
\name{size_effect}
\alias{size_effect}
\title{size effect for pqn}
\usage{
size_effect(x, xref, robust = T)
}
\arguments{
\item{x}{metabolomics profile}

\item{xref}{reference profile}

\item{robust}{TRUE/FALSE}
}
\description{
size effect for pqn
}
\author{
Ettore Mosca
}
