% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitQC.R
\name{splitQC}
\alias{splitQC}
\title{Split the QC samples}
\usage{
splitQC(m_list)
}
\arguments{
\item{metadata}{split in metadata for QC and metadata for samples}

\item{calculate}{mean for each biological replicates}
}
\description{
Split the QC samples
}
