% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{user has to Choose between T-test, U-test, Anova or Kruskal}
\usage{
univariate(
  m_list,
  dirout = "./",
  test_method = c("ttest", "anova", "Utest", "kruskal"),
  paired = c("FALSE", "TRUE"),
  group_factor = "class"
)
}
\arguments{
\item{m_list$data}{(dataframe of samples only biological replicates, technical replicates were collapsed before)}
}
\description{
user has to Choose between T-test, U-test, Anova or Kruskal
}
