% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Univariate analysis}
\usage{
univariate(
  mRList,
  dirout = "./",
  test_method = c("ttest", "Utest", "anova", "kruskal"),
  paired = c("FALSE", "TRUE"),
  group_factor = "class"
)
}
\arguments{
\item{mRList}{dataframe with data of samples that contains only biological replicates. Technical replicates have to be collapsed before using collapse_tech_rep function}

\item{dirout}{output directory}

\item{test_method}{choice of the test to apply}

\item{paired}{TRUE if data are paired, FALSE for unpaired data}

\item{group_factor}{specify the groups to compare}
}
\value{
mRList object with mRList$"testchosen" with univariate analysis
}
\description{
User can choose between T-test, U-test, Anova test  or Kruskal-Wallis test.
}
\examples{
##library(dataset.margheRita)
##dataset(norm_pos)
mRList<-univariate(mRList, test_method="anova",paired=FALSE,group_factor="class")
}
