% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_1spectra.R
\name{visualize_1spectra}
\alias{visualize_1spectra}
\title{Visualize the spectra of good candidates}
\usage{
visualize_1spectra(
  spectra_top = NULL,
  spectra_bottom = NULL,
  name_top = NULL,
  name_bottom = NULL,
  out_file = NULL,
  type = c("mirrored", "overlapped"),
  peak_number = FALSE
)
}
\arguments{
\item{spectra_top}{MS/MS spectra as two-columns data.frame}

\item{spectra_bottom}{MS/MS spectra as two-columns data.frame}

\item{name_top}{name for top spectra}

\item{out_file}{output file}

\item{type}{mirrored or overlapped visualization}

\item{peak_number}{whether to show the peak number or not}

\item{name_bottomname}{for bottom spectra}

\item{peak_id}{peak id}
}
\description{
Visualize the spectra of good candidates
}
