% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_spectra.R, R/visualize_spectra.z.R
\name{visualize_spectra}
\alias{visualize_spectra}
\title{Visualize the spectra of good candidates}
\usage{
visualize_spectra(annotation, RI_lib, RI_sample)

visualize_spectra(annotation, RI_lib, RI_sample)
}
\arguments{
\item{candidate_list}{annotation_list}

\item{x}{one or more metabolites to visualize}
}
\description{
Visualize the spectra of good candidates

visualize the spectra of good candidates
}
