% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_spectra2.R
\name{visualize_spectra2}
\alias{visualize_spectra2}
\title{visualize the spectra of good candidates}
\usage{
visualize_spectra2(x = NULL, RI_sample = NULL, RI_lib = NULL)
}
\description{
visualize the spectra of good candidates
}
